tools/harvester/requirements.txt
httpx>=0.27,<1.0
psycopg[binary]>=3.1,<4.0
boto3>=1.34,<2.0
Pillow>=10.0,<12.0
click>=8.1,<9.0
rich>=13.0,<14.0
